added projects policy

+++ app/Http/Controllers/Admin/ProjectController.php
@@ -26,11 +26,15 @@ class ProjectController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Project::class);
+
         return view('projects.create', ['clients' => Client::all()]);
     }
 
     public function store(Project $project, ValidateProjectRequest $request, CreateModel $createModel)
     {
+        $this->authorize('create', Project::class);
+
         $createModel->handle(Project::class, $request);
 
         return redirect()->route('projects.index');
@@ -38,6 +42,8 @@ class ProjectController extends Controller
 
     public function edit(Project $project)
     {
+        $this->authorize('update', Project::class);
+
         return view('projects.edit', [
             'project' => $project,
             'clients' => Client::all()
@@ -46,6 +52,8 @@ class ProjectController extends Controller
 
     public function update(Project $project, ValidateProjectRequest $request, UpdateModel $updateModel)
     {
+        $this->authorize('update', Project::class);
+
         $updateModel->handle($project, $request);
 
         return redirect()->route('projects.index');
@@ -53,6 +61,8 @@ class ProjectController extends Controller
 
     public function destroy(Project $project)
     {
+        $this->authorize('delete', Project::class);
+
         $project->delete();
 
         return redirect()->route('projects.index');
